refactor: rename app entrypoint
